Remove broken explorer.py
It was giving 500 errors when selected. 

There are several problems with it causing different tracebacks. I think this was an experiment that was not used. I suggest we pull this theme instead of trying to fix it. (ps it's currently renamed on the server to fix it for current users)

If users select this theme they can no longer access the wiki when signed in. http://answers.ros.org/question/209546/cannot-access-wiki-after-changing-theme/?comment=209719#comment-209719

I verified I could not access the wiki with this theme and removing it allows users who have this theme selected to fall back to the default theme (rostheme)