Extend 'Continuous Integration' badge with dev job test stats.
Instead of checking for just the existence of a devel job for a
repository, display some statistics about the latest build (on the badge
itself) and show build history in a drop-down.

The following colours are used:

 - red: repository has failing tests
 - grey: repository has skipped tests
 - green: all tests passed

The ROS wiki PackageHeader macro is extended with logic to load and parse
additional yaml files that contain summaries of the Jenkins build history.
These files are generated and copied after every run of the devel job for
a particular repository.

Packages that do not have devel jobs will not get a 'Continuous
Integration' badge (this is the current default behaviour that has not
changed).

Packages for which no test statistics yaml file can be found (but do have
a devel job registered) will display the regular badge (no test stats nor
history).

Packages for which there is no build history available will display only
the statistics of the latest build.

Finally, packages with devel jobs and build history but with no tests run
during any of those builds will show '0/0' on everything.

--

The authors of this contribution were supported by the European Union's
Horizon 2020 research and innovation programme under grant agreement No
732287 (ROSIN).