Added a warning message on pages that may be outdated.
When a wiki page was not updated recently, a warning message
is displayed in the header, informing the reader that the
information on the wiki page may be outdated.
The default threshold for the warning to appear is 365 days,
but can be adjusted by adding
`outdated_page_warning_threshold = <days>`
to the config file.